requirements.txt
langchain 
langchain_community
langchain_core
pypdfium2
librosa
pyyaml
python-dotenv
faiss-cpu
ctransformers
transformers
streamlit-mic-recorder
InstructorEmbedding
streamlit
chromadb
torch
sentence-transformers==2.3.1
llama-cpp-python